fix(utils): keep set_zero_in_cmap_to_transparent public
The utils split (#696) relocated the public set_zero_in_cmap_to_transparent
into the private _color module, breaking the only public import path:
`from spatialdata_plot.pl.utils import set_zero_in_cmap_to_transparent`.

Move it back to utils.py — a public symbol belongs in a public module, and
a re-export shim would re-introduce the _color<->utils cycle the split
removed (utils is the import-graph leaf). Point the test at the public path.
